pull analytics in pagination component to its own hook
